Modified the project strucure and updated the README with more info about the lib. (#7)
* Ref #3327. Added, documented and tested type: int8.

* #3328 Refactor headers

* Ref #3328. Refactor sources.

* #3328. Added documentation.

* Changes for accepting Pull Request #7

* Changes (2) for accepting Pull Request #7
