+++ .gitignore
@@ -1,3 +1,5 @@
 build
 examples/performance_cmp_versions
 .vscode
+.color_coded
+.ycm_extra_conf.py
